fix broken test and behavior in LayerResource
testIndirectUpdateSRS was broken due to attribute encoding.
The problem was that TaskResource was using LayerResource to
copy the PUT'd layer/featuretype info resulting in an invalid
catalog state. Since the only supported change is SRS, remove
the call to CatalogBuilder that copies the ResourceInfo.

Also remove old, superfluous latLonBoundBox computation and
let this be handled elsewhere. Added assertion to the test to
verify this works.

Minor imports cleanup.

Conflicts:

	src/community/importer/rest/src/test/java/org/geoserver/importer/rest/RESTDataTest.java