[GEOS-6683] Type migrated GeoJSON
I migrated the type from GeoJSON back to an Integer value instead of a
String, this is in line with the new approach that Andrea and Justin
advised me to do. Also I removed BigInt, no idea why IntelliJ put it
there.

Thank you all for your patience with me!