[GEOS-6688] Addition to support more interpolation types.
The nearest neighbor interpolation type was not working because of an issue regarding spelling of the word neighbor. When enforcing one style of spelling we potentially break a lot of existing configurations in GeoServer. That's why it is more convenient to check with startsWith() instead of equalsIngoreCase(). This way all of the existing interpolation methods are working while maintaining compatibility with the differences in spelling.