Moved data store flush on feature type delete to event listener
Moved data store flush from restconfig to ResourcePool to
account for other sources of that event.
Avoid calling DataStore.dispose() as other threads may be using it and has proved to
result in unpredictable errors. Instead, we're calling the datastore
feature type's ContentState.flush() method which forces re-loading
the native type when next used if dealing with a ContentDataStore subclass.

Force removing the cached AttributeTypeInfo for a given FeatureTypeInfo
when the feature type is evicted from the feature type cache, which can
happen by cache cleanup as well as when explicitly removing a feature type.