[GEOS-5391] GetCapabilities 1.1.1 and 1.3.0 does not include GetMap formats for exceptions
Added the missing application/vnd.ogc.se_blank exception that was missing from the valid exceptions for WMS 1.1.1. This exception has been added to the GetCapabilities, but also produces a valid, blank image when an exception is thrown. Also added tests for both the Capbilities and the exception handling, where the exception is compared to a blank image.