Added clip rectangle buffer and generalization factor to use Target CRS
a) minor documentation additions
b) Generalization was being done after transforming to the target CRS, but was using the source CRS generalization factor
c) Changed clipping so the clipping rectangle is 12px bigger than the actual tile boundary. This is what the streaming renderer does. This was done because it hides the introduced clip lines to polygons.
d) changed the clipper to use an envelope instead of a polygon and use the GeometryClipper with some support for dealing with topology exceptions.

All of these are fairly minor.